nutrition-agent/requirements.txt
langchain
langchain-community
langchain-openai
duckduckgo-search
pydantic
requests
python-dotenv
